breaking planechanse and archenemy into smaller components
